KnownLayers arfe no updated in the background. Test not working yet
