+++ src/main/java/org/auscope/portal/core/server/controllers/KnownLayerController.java
@@ -1,10 +1,8 @@
 package org.auscope.portal.core.server.controllers;
 
-import org.auscope.portal.core.services.GoogleCloudMonitoringCachedService;
 import org.auscope.portal.core.services.KnownLayerService;
 import org.auscope.portal.core.view.ViewCSWRecordFactory;
 import org.auscope.portal.core.view.ViewKnownLayerFactory;
-import org.auscope.portal.core.view.knownlayer.KnownLayerGrouping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,18 +20,12 @@ public class KnownLayerController extends BaseCSWController {
     /** Used for requesting groupings of CSWRecords under known layers */
     private KnownLayerService knownLayerService;
 
-    private GoogleCloudMonitoringCachedService stackDriverCachedService = null;
-
     @Autowired
-    public KnownLayerController(KnownLayerService knownLayerService,
-            ViewKnownLayerFactory viewFactory, ViewCSWRecordFactory viewCSWRecordFactory) {
-        super(viewCSWRecordFactory, viewFactory);
+    public KnownLayerController(KnownLayerService knownLayerService, ViewCSWRecordFactory viewCSWRecordFactory) {
+        super(viewCSWRecordFactory);
         this.knownLayerService = knownLayerService;
     }
-    @Autowired(required = false)
-    public void setCachedService(GoogleCloudMonitoringCachedService service) {
-        this.stackDriverCachedService = service;
-    }
+    
     /**
      * Gets a JSON response which contains the representations of each and every "KnownFeatureTypeDefinition".
      *
@@ -42,26 +34,33 @@ public class KnownLayerController extends BaseCSWController {
      * @return
      */
     @RequestMapping("getKnownLayers.do")
-    public ModelAndView getKnownLayers() {
-        KnownLayerGrouping grouping = knownLayerService.groupKnownLayerRecords();
-        if (stackDriverCachedService != null) {
-            return generateKnownLayerResponse(grouping.getKnownLayers(), stackDriverCachedService);
-        } else {
-            return generateKnownLayerResponse(grouping.getKnownLayers());
-        }
+    public ModelAndView getKnownLayers() { 
+        return generateKnownLayerResponse();
     }
-
+  
     /**
-     * Gets a JSON response which contains the representations of each and every "KnownFeatureTypeDefinition".
+     * Utility for generating a response model that represents a number of KnownLayerAndRecord objects. Also adds
+     * service failure information from nagios to the layers (if recorded in the known layer).
      *
-     * Each KnownFeatureTypeDefinition will map [0, N] CSWRecords with display information.
-     *
+     * @param knownLayers
+     *            The known layers to transform
      * @return
      */
-    @RequestMapping("getUnmappedCSWRecords.do")
-    public ModelAndView getUnmappedCSWRecords() {
-        KnownLayerGrouping grouping = knownLayerService.groupKnownLayerRecords();
-
-        return generateCSWRecordResponse(grouping.getUnmappedRecords());
+    protected ModelAndView generateKnownLayerResponse() {
+        return generateJSONResponseMAV(true, knownLayerService.getKnownLayersCache(), "");            
     }
+
+//    /**
+//     * Gets a JSON response which contains the representations of each and every "KnownFeatureTypeDefinition".
+//     *
+//     * Each KnownFeatureTypeDefinition will map [0, N] CSWRecords with display information.
+//     *
+//     * @return
+//     */
+//    @RequestMapping("getUnmappedCSWRecords.do")
+//    public ModelAndView getUnmappedCSWRecords() {
+//        KnownLayerGrouping grouping = knownLayerService.groupKnownLayerRecords();
+//
+//        return generateCSWRecordResponse(grouping.getUnmappedRecords());
+//    }
 }
